refactor: パッケージ名をwebhookからexternalに変更
